requirements.txt
fastapi[standard]
python-dotenv
httpx
langextract[openai]
httpie